Support all HTTP methods and JSON payload
Before this commit, only GET method with an empty payload (ie. body) was supported due to the hardcoded strings in authentication hash computation.

This commit allow to use any supported HTTP method and compute hash with body too.

Please note that due to the usage of faraday here and the lack of methods in it to retrieve the effective URI with query string and the computed body after middlewares processes, we have to compute them ourselve and pry that will match.
In other words, using `JSON.generate(data)` and `URI.encode_www_form(params)` to compute body and URI respectively is working now but could fail if Faraday decides to change its JSON serialization or the query string encoding